# src/main/java/com/example/productapi/dto/ErrorResponse.java
package com.example.productapi.dto;


import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.Builder;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ErrorResponse extends RuntimeException {
    private int status;
    private String error;
    private String message;
}